Change return types from BaseProperty to PropertyWrapper
Change CurrentProperty and LastVisitedProperty return types from BaseProperty?
to PropertyWrapper? instead of casting down when not needed, and provide
modders with full property functionality. Might just end up removing BaseProperty in the future, but I think it might come in handy in the case where we might want to access features that are specific to a single property.